bash master

+++ run.sh
@@ -9,20 +9,60 @@ echo "Usage: ./run.sh docker-swarm dev -- create the dev environment in the dock
 echo "       ./run.sh blueboard staging -- create the destagingv environment in theblueboard repo using blueboard-staging.env config file"
 echo ""
 
-# GITHUB_ACCESS_TOKEN is hardcoded in git.env
+PS3="Select your repo: "
+select repo in blueboard docker other quit
+do
+    case $repo in
+        "blueboard")
+            export GITHUB_REPO="blueboard"
+            break;;
+        "docker")
+            export GITHUB_REPO="docker-swarm"
+            break;;
+        "other")
+            export GITHUB_REPO="foo"
+            break;;
+        "quit")
+            echo "Goodbye..."
+            break;;
+        *)
+           echo "Entry was not recognized";;
+    esac
+done
 
-GITHUB_REPO=${1:-blueboard}
-ENVIRONMENT=${2:-dev}
+PS3="Select your env: "
+select env in dev staging prod
+do
+    case $env in
+        "prod")
+            export ENVIRONMENT="prod"
+            break;;
+        "staging")
+            export ENVIRONMENT="staging"
+            break;;
+        *)
+            export ENVIRONMENT="dev"
+            break;;
+    esac
+done
 
-read -p "Push variables to Github environment ${ENVIRONMENT} to the ${GITHUB_REPO} GitHub repo?  Are you sure? " -n 1 -r
+FILE=${GITHUB_REPO}-${ENVIRONMENT}.env
+if [ ! -f ${FILE} ]; then
+    echo "Environment file named \"${FILE}\" was not found!"
+    exit 1
+fi
+
+echo ""
+read -p "Push variables to Github environment ${ENVIRONMENT} to the ${GITHUB_REPO} GitHub repo from the file ${FILE}?  Are you sure? " -n 1 -r
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]; then
-    echo "Creating a virtual environment"
+
+    # echo "Creating a virtual environment"
     python3 -m venv venv
-    echo "Activating the virtual environment"
+    # echo "Activating the virtual environment"
     source ./venv/bin/activate
-    echo "Installing requirements"
+    # echo "Installing requirements"
     pip3 install -r requirements.txt
-    echo "Creating/Updating environments/repository secrets"
+    # echo "Creating/Updating environments/repository secrets"
     ENVIRONMENT=${ENVIRONMENT} GITHUB_REPO=${GITHUB_REPO} python3 main.py
 fi
